+++ src/main/java/org/apache/xpath/functions/datetime/FuncAdjustTimeToTimezone.java
@@ -111,14 +111,16 @@ public class FuncAdjustTimeToTimezone extends FunctionMultiArgs {
 		if (arg1Timezone != null) {
 		   int tzHrs = arg1Timezone.hours();
 		   int tzMins = arg1Timezone.minutes();
-		   int totalTzHrs = (tzHrs + (tzMins / 60));
+		   double totalTzHrs = (tzHrs + ((double)tzMins / 60)); 
+		   
 		   totalTzHrs = (arg1Timezone.negative() ? (-1 * totalTzHrs) : totalTzHrs);
+		   
 		   if ((totalTzHrs < -14) || (totalTzHrs > 14)) {
 			   throw new javax.xml.transform.TransformerException("FODT0003 : An XPath function adjust-time-to-timezone's "
 							  		                                                                         + "second argument doesn't represent a timezone "
 							  		                                                                         + "value within valid duration range. Timezone value "
 							  		                                                                         + "can be within the range -PT14H and PT14H.", srcLocator);
-		   } 
+		   }
 		}
 				
 		XSDuration arg0XsTimezone = arg0XsTime.getTimezone();
